# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: API for bank cards, users and transfers management
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/cards:
    post:
      tags:
        - Cards
      summary: Create new card
      description: Create a new bank card. Admin only
      operationId: createCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequestDto'
      responses:
        '200':
          description: Card created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto_CardResponseDto'
        '400':
          description: Invalid card data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

    get:
      tags:
        - Cards
      summary: Get user's cards
      description: Get paginated list of user's cards with filtering options
      operationId: getMyCards
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Card status
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: minBalance
          in: query
          description: Minimum balance
          schema:
            type: number
            format: decimal
            minimum: 0
        - name: maxBalance
          in: query
          description: Maximum balance
          schema:
            type: number
            format: decimal
            minimum: 0
        - name: expiryFrom
          in: query
          description: Expiry date from
          schema:
            type: string
            format: date
        - name: expiryTo
          in: query
          description: Expiry date to
          schema:
            type: string
            format: date
        - name: cardholderContains
          in: query
          description: Part of the cardholder's name to search for
          schema:
            type: string
        - name: page
          in: query
          description: Page number (starts with 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            default: createdAt,desc
      responses:
        '200':
          description: Cards retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto_PageCardResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/v1/cards/{cardId}/block:
    post:
      tags:
        - Cards
      summary: Block card
      description: |
        Block a card. For users - creates block request.
        For admins - immediate block.
      operationId: blockCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: |
            USER: block request has been created
            ADMIN: card has been blocked
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiResponseDto_CardResponseDto'
                  - $ref: '#/components/schemas/ApiResponseDto_BlockRequestResponseDto'
        '400':
          description: Invalid card ID or card has been blocked already
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/v1/cards/{cardId}/approve-block:
    post:
      tags:
        - Cards
      summary: Approve card block request
      description: Admin approval for card block request. Admin only
      operationId: approveBlockCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            format: uuid
        - name: username
          in: query
          description: Username (optional)
          schema:
            type: string
      responses:
        '200':
          description: Card block approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto_CardResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card or block request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/v1/cards/{cardId}/reject-block:
    post:
      tags:
        - Cards
      summary: Reject card block request
      description: Admin rejection for card block request. Admin only
      operationId: rejectBlockCard
      security:
        - bearerAuth: [ ]
      parameters:
        - name: cardId
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            format: uuid
        - name: username
          in: query
          description: Username (optional)
          schema:
            type: string
      responses:
        '200':
          description: Card block rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto_CardResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card or block request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/v1/cards/{cardId}/activate:
    post:
      tags:
        - Cards
      summary: Activate card
      description: Activate a previously blocked card. Admin only
      operationId: activateCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto_CardResponseDto'
        '400':
          description: Card is not blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/v1/cards/{cardId}/balance:
    get:
      tags:
        - Cards
      summary: Get card balance
      description: Get current balance of a specific card
      operationId: getBalance
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto_BigDecimal'
        '400':
          description: Invalid card ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - user doesn't have access to this card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token format: `Bearer <token>`
        Token may be retrieved `/api/v1/auth/login`
  schemas:
    CardStatus:
      type: string
      enum: [ACTIVE, BLOCKED, EXPIRED, PENDING_BLOCK]
      description: Bank card status
      example: ACTIVE
...
    CardRequestDto:
      type: object
      required: [cardholder, cardNumber, expiryDate]
      properties:
        cardholder:
          type: string
          description: Cardholder name
          minLength: 2
          maxLength: 100
          example: Roman Kuzmich
        cardNumber:
          type: string
          description: Card number (16 digits)
          pattern: '^[0-9]{16}$'
          example: '4111111111111111'
        expiryDate:
          type: string
          format: date
          description: Expiry date
          example: '2026-12-31'
...
    CardResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique card ID
          example: 550e8400-e29b-41d4-a716-446655440000
        maskedNumber:
          type: string
          description: Masked card number
          example: '**** **** **** 1234'
        cardholder:
          type: string
          description: Cardholder name
          example: Roman Kuzmich
        expiryDate:
          type: string
          format: date
          description: Expiry date
          example: '2026-12-31'
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          format: decimal
          description: Current card balance
          example: 1500.75
        createdAt:
          type: string
          format: date-time
          description: Card creation timestamp
          example: '2024-01-15T10:30:00'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2024-01-15T14:45:00'
    BlockRequestResponseDto:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
          description: Unique card ID
          example: 550e8400-e29b-41d4-a716-446655440000
        cardStatus:
          $ref: '#/components/schemas/CardStatus'
        hasPendingRequest:
          type: boolean
          description: Pending block request presence
          example: true
        requestedAt:
          type: string
          format: date-time
          description: Block request creation timestamp
          example: '2024-01-15T14:30:00'
        requestedBy:
          type: string
          description: Block request creator username
          example: john_doe
        message:
          type: string
          description: Information message
          example: Card block request has been sent to administrator for review
...
    ApiResponseDto:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success flag
          example: true
        message:
          type: string
          description: Message text
          example: Operation completed successfully
        data:
          type: object
          description: Response data (may be null in case of error)
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: '2024-01-15T14:30:00'
    ErrorResponseDto:
      type: object
      properties:
        success:
          type: boolean
          default: false
          description: Operation success flag
          example: false
        message:
          type: string
          description: Main error message
          example: Validation failed
        errors:
          type: array
          description: Detailed errors message (may be null)
          items:
            type: string
          example: ["Email must be valid", "Password is required"]
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T14:30:00'
...
    ApiResponseDto_CardResponseDto:
      allOf:
        - $ref: '#/components/schemas/ApiResponseDto'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CardResponseDto'
    ApiResponseDto_BlockRequestResponseDto:
      allOf:
        - $ref: '#/components/schemas/ApiResponseDto'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BlockRequestResponseDto'
...
    ApiResponseDto_BigDecimal:
      allOf:
        - $ref: '#/components/schemas/ApiResponseDto'
        - type: object
          properties:
            data:
              type: number
              format: decimal
              example: 1500.75
...
    Page:
      type: object
      properties:
        content:
          type: array
          description: Page content
        pageable:
          type: object
          properties:
            pageNumber:
              type: integer
              description: Page number
            pageSize:
              type: integer
              description: Page size
            sort:
              type: object
              properties:
                sorted:
                  type: boolean
                unsorted:
                  type: boolean
                empty:
                  type: boolean
            offset:
              type: integer
            paged:
              type: boolean
            unpaged:
              type: boolean
        totalPages:
          type: integer
          description: Total page count
        totalElements:
          type: integer
          description: Total elements count
        last:
          type: boolean
          description: Is this the last page
        size:
          type: integer
          description: Page size
        number:
          type: integer
          description: Page number
        sort:
          type: object
          properties:
            sorted:
              type: boolean
            unsorted:
              type: boolean
            empty:
              type: boolean
        numberOfElements:
          type: integer
          description: Current page elements count
        first:
          type: boolean
          description: Is this the first page
        empty:
          type: boolean
          description: Is the page empty
    PageCardResponseDto:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/CardResponseDto'
...
    ApiResponseDto_PageCardResponseDto:
      allOf:
        - $ref: '#/components/schemas/ApiResponseDto'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageCardResponseDto'
...